requirements.txt
geopandas>=0.10.0
shapely>=1.8.0
pyproj>=3.0.0
pyyaml>=5.4.0
matplotlib>=3.5.0
